dataengineering/challenge-3/event-streaming/requirements.txt
kafka-python==2.0.2
requests
requests_oauthlib